Refactor module structure: move error.rs to profile/ and git_config.rs to src/
- Move src/error.rs to src/profile/error.rs to group errors with profile operations
- Move src/profile/git_config.rs to src/git_config.rs for better separation of concerns
- Update all import statements across affected files to use new module paths
- All tests continue to pass, no functional changes
