+++ src/profile.rs
@@ -1,4 +1,4 @@
-pub mod git_config;
+pub mod error;
 pub mod git_config_git2;
 pub mod git_profile_dir;
 pub mod list;
